# openapi.yaml
openapi: 3.1.0
info:
  title: Orchestrix API
  version: 1.0.0
  description: |
    API de orquestração de workflows e execuções.

    ## Autenticação
    Todas as rotas (exceto health checks) requerem autenticação via JWT Bearer token.

    ## Multi-tenancy
    Todos os recursos são isolados por tenant. O tenant é determinado pelo token JWT.

    ## Paginação
    Endpoints de listagem suportam paginação via query parameters `page` e `limit`.
  contact:
    name: Orchestrix Team
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development
  - url: https://api.orchestrix.io/v1
    description: Production
security:
  - BearerAuth: []
...
paths:
...
  /workflows:
    get:
      operationId: listWorkflows
      summary: List workflows
      description: Returns a paginated list of workflows for the current tenant
      tags:
        - Workflows
      parameters:
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/LimitQuery'
        - $ref: '#/components/parameters/WorkflowStatusQuery'
        - $ref: '#/components/parameters/SearchQuery'
      responses:
        '200':
          description: List of workflows
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedWorkflows'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      operationId: createWorkflow
      summary: Create workflow
      description: Creates a new workflow definition
      tags:
        - Workflows
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkflowRequest'
            examples:
              simple:
                summary: Simple workflow
                value:
                  name: Order Processing
                  description: Processes incoming orders
                  definition:
                    steps:
                      - name: validate
                        type: validate_order
                      - name: process
                        type: process_payment
              scheduled:
                summary: Scheduled workflow
                value:
                  name: Daily Report
                  definition:
                    steps:
                      - name: generate
                        type: generate_report
                  schedule: 0 9 * * *
      responses:
        '201':
          description: Workflow created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /workflows/{id}:
    get:
      operationId: getWorkflow
      summary: Get workflow
      description: Returns a single workflow by ID
      tags:
        - Workflows
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Workflow found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      operationId: updateWorkflow
      summary: Update workflow
      description: Updates an existing workflow. Version is auto-incremented.
      tags:
        - Workflows
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkflowRequest'
            examples:
              update_status:
                summary: Activate workflow
                value:
                  status: active
              update_definition:
                summary: Update definition
                value:
                  definition:
                    steps:
                      - name: step1
                        type: action
      responses:
        '200':
          description: Workflow updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      operationId: deleteWorkflow
      summary: Delete workflow
      description: Deletes a workflow by ID
      tags:
        - Workflows
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /workflows/{id}/execute:
    post:
      operationId: executeWorkflow
      summary: Execute workflow
      description: Triggers a new execution of the workflow
      tags:
        - Workflows
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecuteWorkflowRequest'
            examples:
              with_input:
                summary: With input data
                value:
                  input:
                    order_id: '12345'
                    customer_id: '67890'
              empty:
                summary: No input
                value: {}
      responses:
        '202':
          description: Execution started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /workflows/{id}/executions:
    get:
      operationId: listWorkflowExecutions
      summary: List workflow executions
      description: Returns a paginated list of executions for a specific workflow
      tags:
        - Workflows
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/LimitQuery'
        - $ref: '#/components/parameters/ExecutionStatusQuery'
      responses:
        '200':
          description: List of executions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedExecutions'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /executions:
    get:
      operationId: listExecutions
      summary: List executions
      description: Returns a paginated list of all executions for the current tenant
      tags:
        - Executions
      parameters:
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/LimitQuery'
        - $ref: '#/components/parameters/ExecutionStatusQuery'
      responses:
        '200':
          description: List of executions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedExecutions'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /executions/{id}:
    get:
      operationId: getExecution
      summary: Get execution
      description: Returns a single execution by ID
      tags:
        - Executions
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Execution found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Keycloak
  schemas:
...
    WorkflowStatus:
      type: string
      enum:
        - active
        - inactive
        - draft
      description: Workflow status
    Workflow:
      type: object
      description: Workflow definition for orchestration
      required:
        - id
        - tenant_id
        - name
        - definition
        - status
        - version
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique workflow identifier
          examples:
            - 550e8400-e29b-41d4-a716-446655440000
        tenant_id:
          type: string
          format: uuid
          description: Tenant identifier
        name:
          type: string
          minLength: 3
          maxLength: 100
          description: Workflow name
          examples:
            - Order Processing Workflow
        description:
          type: string
          maxLength: 500
          description: Optional workflow description
          examples:
            - Processes incoming orders and triggers notifications
        definition:
          type: object
          additionalProperties: true
          description: Workflow definition as JSON
        schedule:
          type: string
          description: Cron expression for scheduled execution (e.g., "0 * * * *" or "@daily")
          examples:
            - 0 * * * *
            - '@daily'
        status:
          $ref: '#/components/schemas/WorkflowStatus'
        version:
          type: integer
          minimum: 1
          description: Version number (auto-incremented on update)
          examples:
            - 1
        created_by:
          type: string
          format: uuid
          description: User who created the workflow
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
    PaginatedWorkflows:
      type: object
      description: Paginated list of workflows
      required:
        - data
        - total
        - page
        - limit
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Workflow'
        total:
          type: integer
          minimum: 0
          description: Total number of items
        page:
          type: integer
          minimum: 1
          description: Current page
        limit:
          type: integer
          minimum: 1
          maximum: 100
          description: Items per page
    Error:
      type: object
      description: Error response
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Error code
          examples:
            - VALIDATION_ERROR
            - NOT_FOUND
            - UNAUTHORIZED
            - INTERNAL_ERROR
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          additionalProperties: true
          description: Additional error details
    CreateWorkflowRequest:
      type: object
      description: Request body for creating a workflow
      required:
        - name
        - definition
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 100
          description: Workflow name
        description:
          type: string
          maxLength: 500
          description: Optional description
        definition:
          type: object
          additionalProperties: true
          description: Workflow definition
        schedule:
          type: string
          description: Optional cron schedule
    WorkflowResponse:
      type: object
      description: Single workflow response wrapper
      required:
        - data
      properties:
        data:
          $ref: '#/components/schemas/Workflow'
    UpdateWorkflowRequest:
      type: object
      description: Request body for updating a workflow
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 100
        description:
          type: string
          maxLength: 500
        definition:
          type: object
          additionalProperties: true
        schedule:
          type: string
        status:
          $ref: '#/components/schemas/WorkflowStatus'
    ExecuteWorkflowRequest:
      type: object
      description: Request body for executing a workflow
      properties:
        input:
          type: object
          additionalProperties: true
          description: Input data for the workflow execution
    ExecutionStatus:
      type: string
      enum:
        - pending
        - running
        - completed
        - failed
        - cancelled
      description: Execution status
    Execution:
      type: object
      description: Workflow execution instance
      required:
        - id
        - tenant_id
        - workflow_id
        - status
        - created_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique execution identifier
        tenant_id:
          type: string
          format: uuid
          description: Tenant identifier
        workflow_id:
          type: string
          format: uuid
          description: Parent workflow identifier
        temporal_workflow_id:
          type: string
          description: Temporal workflow ID
        temporal_run_id:
          type: string
          description: Temporal run ID
        status:
          $ref: '#/components/schemas/ExecutionStatus'
        input:
          type: object
          additionalProperties: true
          description: Input data provided at execution
        output:
          type: object
          additionalProperties: true
          description: Output data from execution
        error:
          type: string
          description: Error message if execution failed
        started_at:
          type: string
          format: date-time
          description: Execution start time
        completed_at:
          type: string
          format: date-time
          description: Execution completion time
        created_by:
          type: string
          format: uuid
          description: User who triggered the execution
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
    ExecutionResponse:
      type: object
      description: Single execution response wrapper
      required:
        - data
      properties:
        data:
          $ref: '#/components/schemas/Execution'
    PaginatedExecutions:
      type: object
      description: Paginated list of executions
      required:
        - data
        - total
        - page
        - limit
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Execution'
        total:
          type: integer
          minimum: 0
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
...
  responses:
    Unauthorized:
      description: Unauthorized - Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            missing_token:
              summary: Missing Token
              value:
                code: UNAUTHORIZED
                message: Missing or invalid authorization header
            expired_token:
              summary: Expired Token
              value:
                code: UNAUTHORIZED
                message: Token has expired
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            internal_error:
              summary: Internal Error
              value:
                code: INTERNAL_ERROR
                message: An unexpected error occurred
    BadRequest:
      description: Bad Request - Invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            validation_error:
              summary: Validation Error
              value:
                code: VALIDATION_ERROR
                message: Invalid request body
                details:
                  name: must be at least 3 characters
    Conflict:
      description: Conflict - Resource already exists or state conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            conflict:
              summary: Conflict
              value:
                code: CONFLICT
                message: Resource already exists
    NotFound:
      description: Not Found - Resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            not_found:
              summary: Not Found
              value:
                code: NOT_FOUND
                message: Resource not found
    NoContent:
      description: No Content - Successful operation with no response body
    UnprocessableEntity:
      description: Unprocessable Entity - Business rule violation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            business_error:
              summary: Business Rule Violation
              value:
                code: BUSINESS_ERROR
                message: Cannot execute inactive workflow
...
  parameters:
    PageQuery:
      name: page
      in: query
      required: false
      description: Page number (1-based)
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitQuery:
      name: limit
      in: query
      required: false
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    WorkflowStatusQuery:
      name: status
      in: query
      required: false
      description: Filter by workflow status
      schema:
        type: string
        enum:
          - active
          - inactive
          - draft
    SearchQuery:
      name: search
      in: query
      required: false
      description: Search term
      schema:
        type: string
        maxLength: 100
    IdPath:
      name: id
      in: path
      required: true
      description: Resource UUID
      schema:
        type: string
        format: uuid
    ExecutionStatusQuery:
      name: status
      in: query
      required: false
      description: Filter by execution status
      schema:
        type: string
        enum:
          - pending
          - running
          - completed
          - failed
          - cancelled
...